Correction to signature of PolynomialKernel.setoffset method

+++ dynaml-core/src/main/scala-2.11/io/github/mandar2812/dynaml/kernels/PolynomialKernel.scala
@@ -22,7 +22,7 @@ class PolynomialKernel(
     state += ("degree" -> d.toDouble)
   }
 
-  def setoffset(o: Int): Unit = {
+  def setoffset(o: Double): Unit = {
     this.offset = o
     state += ("offset" -> o)
   }
